test/mock-sysfs: create NHI devices for domains
When creating a domain, also create a corresponding native host
interface PCIe device for it, and use that as the parent.
This better reflects the reality, where domain udev devices are
children of the thunderbolt PCIe device.
The PCI id of the NHI device can be controlled at domain creation
time via the 'nhi' argument. It defaults to '0x15d2', i.e. Alpine
Ridge.