test/integration: simulate native host interface
Add the native host interface (NHI) udev device as a parent of the
thunderbolt domain for both, the device mock tree and the simple
add_host_domain function. The NHI is a PCI device, which bolt will
use to determine certain thunderbolt functionality, i.e. if the
uuid of the controller is stable.