domain: method to see if it can be safely deleted
Add a new method, bolt_domain_can_delete, which can be used to
check if a domain can safely be deleted without the loss of data.
Currently, the only case where this can happen if the boot acl
journal for the domain has entries.
The function returns FALSE and sets and GError accordingly in
case it is not safe to delete the domain.