test: check for bolt_domain_can_delete
Check that if a domain has journal entries, bolt_domain_can_delete
indeed returns FALSE, and TRUE otherwise.
This code modifies the original test to use the original domain
object 'd1' to delete the domain instead of the object that was
fetched from the store. This is so that there are not two objects
referring to the domain.