test/integration: cleanup stale domains check
Simulate stale domains by creating a store with four domains, one
of which has a bootacl journal with entries. Since the store does
not have a 'version' file, an upgrade will be performed which will
lead to the cleanup of stale domains. All domains, except the one
with the non-empty bootacl journal should be cleaned up.