store: introduce store versioning
Add version information to the on-disk store and expose that
information via a new "version" store object property. The
on-disk information is stored in a new 'version' file, that
is located at the root of the store. In case no such file is
found, the store version is defined as 0 (zero).
When a store is opened, and the root of the store is completely
empty (i.e. no sub-directories or the version file is present),
the store is assumed to be new and is initialized by creating
the 'version' containing the BOLT_STORE_VERSION number.