store: convert to be initable
Convert the store use the GInitable interface. This prepares for
store on-disk initialization that can fail. Currently, nothing
is actually done in initialize and thus it will never fail.

Adapt all tests to check for potential store creation errors and
also convert the manager's store initialization method to handle
the error.