store: introduce store upgrades
Add a new function to upgrade the store from version "0" to "1".
Except the addition of the 'version' file there is no difference
between the two versions and thus creating said 'version' file
is all the store upgrade does. This can still be useful, because
it can be used as a marker, e.g. the manager can use it to clean
up the store once when upgrading from version 0.