manager: un-export device in deregister_device
Remove some duplicated code by moving the un-export code for the
device to a central location: `manager_deregister_device`. This
method will be called for all devices that can be de-registered,
i.e. either devices that got disconnected and are not store via
the `handle_udev_device_removed` udev event handler, or devices
that got remove from the store and are disconnected. The latter
is handled via `handle_store_device_removed`. In both cases the
device should be de-registered and (if exported) be un-exported.
Thus this is where the un-export code is now moved to.