manager: clean up stale domains on store upgrade
When the store is upgraded, clean up any stale domains. Previous
versions of bolt (< 0.9.1) would store domains that did not have
stable uuids, i.e. their uuid would change on every boot. These
stored domains would be loaded and exported but would never come
online, since they can not ever by matched to an online domain.
When the store cleanup method is executed, in the manager's
initialization method, the domain controller should either be
powered (they use runtime power management only and are thus not
physically powered down on ice lake anyway) or force-powered and
thus any domain that is not yet online will be removed. This can
fail if the domain has a boot acl journal with entries, which
protects the loss of data for domains that are not force-powered,
maybe because they are even slower, but do have a stable uuid.