manager+sysfs: wait for hosts instead of domains
In the manager's initialization, we want to make sure that we have
at least one domain controller online in order detect the security
level. Additionally, now that stale domains are cleaned up on store
upgrades, we also want to make sure that domain controllers are
fully initialized and the corresponding BotlDomain is online, so
that we don't accidentally clean up a non-stale domain. In order
to do so, we wait for at least one host controller to be present,
not just the plain domain. For this we enumerate all domains and
for each domain we enumerate the thunderbolt udev device directly
beneath it (should only be one).
Rename and re-purpose the various `count_domains` functions.
Adapt the test to add host devices to the corresponding domains.