manager: only import hosts with stable uuids
We started auto-importing the host device in commit f9205d1, with
the reason that in non-native enumeration mode, the host device
will disappear and the generation information will thus not be
available anymore. But on embedded thunderbolt controller, the
host device has an unstable unique_id, which means that host device
will have a different uuid on next boot and thus can not be matched
with the previously stored one. Thus stale host devices would
accumulate in the store. Luckily, on embedded thunderbolt, the host
device does not have device/vendor sysfs attributes, which means
that currently creating a BoltDevice for it will not work, which is
a bug, that also prevents stale host devices for now. Once that bug
will get fixed we were to get stale host records, unless we don't
store those host devices. To do so we check if the domain of the
host device is stored which is an indicator of the uuid stability,
checked via the pci id of the native host interface. If the domain
is stored we store the host and vice versa.