manager: don't de-duplicate host device names
When labeling a device, we normally do a duplication check, and
if we find more than once device with the same name, we append
a number suffix. Let's not do that for host devices, since they
are arguably represent the same thing, even if technically the
are different controllers with their own unique id.