sysfs: method to perform host identification
Add a new method to do host identification, which will first check
if the host device can be identified via the normal device ident
method. This will fail in case of embedded thunderbolt controller,
which do not have a DROM and have no device and vendor information
and thus no corresponding sysfs attributes. In that case we fall
back to using the SMBIOS/DMI information for the system to ident
the host.