names: add dmi id sysfs path and attribute names
The SMBIOS/DMI information is exported to userspace via sysfs,
which is convenient, because the information can thus be read
conveniently and as a non-root user.
Define constants for relevant information that can be used to
identify the computer / host, like the product, vendor and the
general path to the sysfs device providing that information.