guard: ensure guard->path is always freed
Normally 'path' gets freed in bolt_guard_remove, which is called
from the dispose method of the BoltGuard. There is one exception
though: when the pipe (named fifo) is still active, we do not
unlink the path, since we want to recover that guard later. In
that case the memory for path will be leaked. Add an additional
g_clear_pointer for path in the 'finalize' method to ensure that
path always gets freed.