test/common: conditionally re-enable atime checks
The checks for modifying the access time via bolt_fs_touch was
previously disabled because it did not work on file-systems
that have the noatime flag set (for very obvious reasons).
Separate these checks out and re-enable them but make their
execution depended on the file-system support for access times,
as reported via statvfs(2).