tests/integration: fix running with systemd 251
udev/systemd, starting version 251 (a791061) started to verify that
paths are actually on sysfs, i.e. the fs type is SYSFS_MAGIC. Since
we use umockdev to emulate sysfs on normal file system this check
will now fail. There is an exit hatch in form of an environment
variable called `SYSTEMD_DEVICE_VERIFY_SYSFS` that when set to `0`
makes systemd/udev forgo that check.  umockdev >= 0.17.10 does set
that variable itself but in Python changes to the env by libraries
does not effect the env of the running script. Thus we have to set
it ourselves.